pyproject.toml
[project]
name = "mcp-server-starrocks"
version = "0.1.0"
description = "official MCP server for StarRocks"
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "mysql-connector-python>=9.2.0",
]

[[project.authors]]
name = "changbinglin"
email = "changbinglin@starrocks.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
mcp-server-starrocks = "mcp_server_starrocks:main"

[project.urls]
Home = "https://github.com/starrocks/mcp-server-starrocks"
